backend/services/newsletter-service/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-environ==0.11.2
django-cors-headers==4.3.1
django-redis==5.4.0
drf-yasg==1.21.7
psycopg2-binary==2.9.9
gunicorn==21.2.0
django-ratelimit==4.1.0
celery==5.3.4
django-celery-beat==2.5.0
requests==2.31.0
html2text==2020.1.16
